+++ apps/interunit-dot-dev/CHANGELOG.md
@@ -1,5 +1,21 @@
 # interunit-dot-dev
 
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @interunit/popover@0.0.26
+  - @interunit/config@0.0.17
+  - @interunit/modal@0.0.20
+  - @interunit/combobox@0.0.7
+  - @interunit/a11y@0.0.22
+  - @interunit/collapsible@0.0.20
+  - @interunit/form@0.0.16
+  - @interunit/primitives@0.0.24
+  - @interunit/responsive@0.0.16
+  - @interunit/tabs@0.0.16
+
 ## 0.1.10
 
 ### Patch Changes