+++ packages/collapsible/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @interunit/collapsible
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @interunit/config@0.0.17
+  - @interunit/a11y@0.0.22
+  - @interunit/primitives@0.0.24
+
 ## 0.0.19
 
 ### Patch Changes