+++ packages/crossplatform/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @interunit/crossplatform
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @interunit/config@0.0.17
+  - @interunit/toolbox@0.0.17
+
 ## 0.0.16
 
 ### Patch Changes