+++ packages/popover/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @interunit/popover
 
+## 0.0.26
+
+### Patch Changes
+
+- Modal asChild adjusments + small fixes
+- Updated dependencies
+  - @interunit/config@0.0.17
+  - @interunit/modal@0.0.20
+  - @interunit/a11y@0.0.22
+  - @interunit/crossplatform@0.0.17
+  - @interunit/primitives@0.0.24
+  - @interunit/toolbox@0.0.17
+
 ## 0.0.25
 
 ### Patch Changes