+++ packages/primitives/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @interunit/primitives
 
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @interunit/config@0.0.17
+  - @interunit/crossplatform@0.0.17
+  - @interunit/toolbox@0.0.17
+
 ## 0.0.23
 
 ### Patch Changes