+++ packages/tabs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @interunit/tabs
 
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @interunit/config@0.0.17
+  - @interunit/a11y@0.0.22
+  - @interunit/primitives@0.0.24
+  - @interunit/toolbox@0.0.17
+
 ## 0.0.15
 
 ### Patch Changes