+++ packages/toolbox/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @interunit/toolbox
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @interunit/config@0.0.17
+
 ## 0.0.16
 
 ### Patch Changes